Improve ritual actions
